@@ -3,8 +3,8 @@ using System.Linq.Expressions;
 using System.Reflection;
 using System.Threading.Tasks;
 using Assertive.Analyzers;
-using Assertive.Expressions;
 using Assertive.Helpers;
+using static Assertive.Expressions.ExpressionHelper;
 
 namespace Assertive
 {
@@ -46,7 +46,7 @@ namespace Assertive
       
       try
       {
-        var task = (Task) expression.Compile(true).DynamicInvoke();
+        var task = (Task) expression.Compile(true).DynamicInvoke()!;
 
         await task;
       }
@@ -55,7 +55,7 @@ namespace Assertive
         if (expectedExceptionType != null && !expectedExceptionType.IsInstanceOfType(ex))
         {
           return ExceptionHelper.GetException(
-            $"Expected {ExpressionHelper.ExpressionToString(bodyExpression)} to throw an exception of type {expectedExceptionType.FullName}, but it threw an exception of type {ex.GetType().FullName} instead.");
+            $"Expected {ExpressionToString(bodyExpression)} to throw an exception of type {expectedExceptionType.FullName}, but it threw an exception of type {ex.GetType().FullName} instead.");
         }
         
         threw = true;
@@ -63,7 +63,7 @@ namespace Assertive
 
       if (!threw)
       {
-        return ExceptionHelper.GetException($"Expected {ExpressionHelper.ExpressionToString(bodyExpression)} to throw an exception, but it did not.");
+        return ExceptionHelper.GetException($"Expected {ExpressionToString(bodyExpression)} to throw an exception, but it did not.");
       }
 
       return null;
@@ -85,7 +85,7 @@ namespace Assertive
         if (expectedExceptionType != null && !expectedExceptionType.IsInstanceOfType(ex.InnerException))
         {
           return ExceptionHelper.GetException(
-            $"Expected {ExpressionHelper.ExpressionToString(bodyExpression)} to throw an exception of type {expectedExceptionType.FullName}, but it threw an exception of type {ex.InnerException.GetType().FullName} instead.");
+            $"Expected {ExpressionToString(bodyExpression)} to throw an exception of type {expectedExceptionType.FullName}, but it threw an exception of type {ex.InnerException!.GetType().FullName} instead.");
         }
         
         threw = true;
@@ -93,7 +93,7 @@ namespace Assertive
 
       if (!threw)
       {
-        return ExceptionHelper.GetException($"Expected {ExpressionHelper.ExpressionToString(bodyExpression)} to throw an exception, but it did not.");
+        return ExceptionHelper.GetException($"Expected {ExpressionToString(bodyExpression)} to throw an exception, but it did not.");
       }
 
       return null;
