@@ -10,12 +10,12 @@ namespace Assertive.Helpers
     {
       var assemblies = AppDomain.CurrentDomain.GetAssemblies();
 
-      var assembly = assemblies.FirstOrDefault(a => a.FullName.StartsWith(assemblyName + ",", StringComparison.OrdinalIgnoreCase));
+      var assembly = assemblies.FirstOrDefault(a => a.FullName?.StartsWith(assemblyName + ",", StringComparison.OrdinalIgnoreCase) == true);
 
       if (assembly == null && assemblyPrefix != null)
       {
         var frameworkLoadedAtAll =
-          assemblies.Any(a => a.FullName.StartsWith(assemblyPrefix, StringComparison.OrdinalIgnoreCase));
+          assemblies.Any(a => a.FullName?.StartsWith(assemblyPrefix, StringComparison.OrdinalIgnoreCase) == true);
 
         // In case of xUnit the assertion DLL is not loaded if you don't use any xUnit assertions
         // so check if we can find any assemblies that match the prefix and if so make an attempt
